Cells Restriction
This is a brief macro to restrict specific cells of the worksheets. The purpose is to allow each cell to display only two lines of characters, and each line's characters should not be more than what's specified. A user is not allowed to type in rebundant lines or rebundant characters in each line, otherwise a window will pop up and the content of the cell will be trimmed.